NES-971 Treates export_responses_by_token and list_groups API methods; updates docstrings; refactors export.export method; PEP-8 adjustments; updates survey.survey_utils methods/docstrings; updates export.views method

+++ patientregistrationsystem/qdc/plugin/tests/test_views.py
@@ -1,5 +1,6 @@
 import csv
 import io
+import os
 import re
 import shutil
 import tempfile
@@ -200,7 +201,10 @@ class PluginTest(ExportTestCase):
         self.assertEqual(message, 'Error: some thing went wrong consuming LimeSurvey API. Please try again. If '
                                   'problem persists please contact System Administrator.')
 
-        shutil.rmtree(self.TEMP_MEDIA_ROOT)
+        # TODO (NES-971): put here and below because if clause does not work when running all class suit.
+        #  Verify why!
+        if os.path.exists(self.TEMP_MEDIA_ROOT):
+            shutil.rmtree(self.TEMP_MEDIA_ROOT)
 
     @override_settings(MEDIA_ROOT=TEMP_MEDIA_ROOT)
     @patch('survey.abc_search_engine.Server')
@@ -220,7 +224,8 @@ class PluginTest(ExportTestCase):
         self.assertEqual(message, 'Error: some thing went wrong consuming LimeSurvey API. Please try again. If '
                                   'problem persists please contact System Administrator.')
 
-        shutil.rmtree(self.TEMP_MEDIA_ROOT)
+        if os.path.exists(self.TEMP_MEDIA_ROOT):
+            shutil.rmtree(self.TEMP_MEDIA_ROOT)
 
     @override_settings(MEDIA_ROOT=TEMP_MEDIA_ROOT)
     @patch('survey.abc_search_engine.Server')
@@ -228,6 +233,49 @@ class PluginTest(ExportTestCase):
         set_limesurvey_api_mocks(mockServer)
         # Could not get responses
         mockServer.return_value.export_responses.side_effect = 2*[{'status': 'No Data, survey table does not exist.'}]
+        self._create_basic_objects()
+        response = self.client.post(
+            reverse('send_to_plugin'),
+            data={
+                'opt_floresta': ['on'], 'patient_selected': ['age*age', 'gender__name*gender'],
+                'patients_selected[]': [str(self.patient.id)]
+            })
+        self.assertRedirects(response, reverse('send_to_plugin'))
+        message = str(list(get_messages(response.wsgi_request))[0])
+        self.assertEqual(message, 'Error: some thing went wrong consuming LimeSurvey API. Please try again. If '
+                                  'problem persists please contact System Administrator.')
+
+        shutil.rmtree(self.TEMP_MEDIA_ROOT)
+
+    @override_settings(MEDIA_ROOT=TEMP_MEDIA_ROOT)
+    @patch('survey.abc_search_engine.Server')
+    def test_POST_send_to_plugin_get_error_in_consuming_limesurvey_api_returns_error_message4(self, mockServer):
+        set_limesurvey_api_mocks(mockServer)
+        # Could not get responses by token
+        mockServer.return_value.export_responses_by_token.side_effect = \
+            4 * [{'status': 'No Data, survey table does not exist.'}]
+        self._create_basic_objects()
+        response = self.client.post(
+            reverse('send_to_plugin'),
+            data={
+                'opt_floresta': ['on'], 'patient_selected': ['age*age', 'gender__name*gender'],
+                'patients_selected[]': [str(self.patient.id)]
+            })
+        self.assertRedirects(response, reverse('send_to_plugin'))
+        message = str(list(get_messages(response.wsgi_request))[0])
+        self.assertEqual(message, 'Error: some thing went wrong consuming LimeSurvey API. Please try again. If '
+                                  'problem persists please contact System Administrator.')
+
+        if os.path.exists(self.TEMP_MEDIA_ROOT):
+            shutil.rmtree(self.TEMP_MEDIA_ROOT)
+
+    @override_settings(MEDIA_ROOT=TEMP_MEDIA_ROOT)
+    @patch('survey.abc_search_engine.Server')
+    def test_POST_send_to_plugin_get_error_in_consuming_limesurvey_api_returns_error_message5(self, mockServer):
+        set_limesurvey_api_mocks(mockServer)
+        # Could not list groups
+        mockServer.return_value.list_groups.side_effect = \
+            8 * [{'status': 'No groups found'}]
         self._create_basic_objects()
         response = self.client.post(
             reverse('send_to_plugin'),
